+++ crates/rspack_core/src/compilation/build_module_graph/mod.rs
@@ -1,6 +1,7 @@
 mod graph_updater;
 mod lazy_barrel_artifact;
 mod module_executor;
+pub mod pass;
 
 use std::sync::atomic::Ordering;
